LayoutTests/http/tests/loading/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-deletion-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Test for Partitioned and Unpartitioned Cookie Deletion
main frame - didChangeLocationWithinPageForFrame
main frame - willPerformClientRedirectToURL: http://localhost:8000/loading/resourceLoadStatistics/resources/set-cookie.php?name=firstPartyCookie&value=value#http://127.0.0.1:8000/loading/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-deletion.html#step2 
main frame - didFinishDocumentLoadForFrame
main frame - didFinishLoadForFrame
main frame - didStartProvisionalLoadForFrame
main frame - didCancelClientRedirectForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - willPerformClientRedirectToURL: http://127.0.0.1:8000/loading/resourceLoadStatistics/partitioned-and-unpartitioned-cookie-deletion.html#step2 
main frame - didStartProvisionalLoadForFrame
main frame - didCancelClientRedirectForFrame
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Test for Partitioned and Unpartitioned Cookie Deletion
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame1-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishDocumentLoadForFrame
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame2-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame2-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame2-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
main frame - didChangeLocationWithinPageForFrame
main frame - didChangeLocationWithinPageForFrame
frame "<!--framePath //<!--frame3-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame4-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame3-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame3-->-->" - didFinishLoadForFrame
frame "<!--framePath //<!--frame4-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame4-->-->" - didFinishDocumentLoadForFrame
  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Setting partitioned, third party cookie.


--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
Should only receive partitioned, third party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'thirdPartyCookie'.
Client-side document.cookie: thirdPartyCookie=value

--------
Frame: '<!--framePath //<!--frame3-->-->'
--------
After removal, should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--framePath //<!--frame4-->-->'
--------
After user interaction, should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Did not receive cookie named 'thirdPartyCookie'.
Client-side document.cookie:
